requirements.txt
allure-behave==2.8.12
allure-python-commons==2.8.12
behave==1.2.6
nose==1.3.7
pylint==2.5.3
selenium==3.141.0